requirements.txt
Django==4.2.2
djangorestframework==3.14.0
python-dotenv==1.0.0
celery==5.3.1
redis==4.5.5
python-telegram-bot==20.3
psycopg2-binary==2.9.6
django-cors-headers==4.1.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
